requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
idna==3.4
kafka-python==2.0.2
oauthlib==3.2.2
pymongo==4.3.2
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.12.0
urllib3==1.26.12
